backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Caching
redis==5.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
bcrypt==4.1.2
firebase-admin==6.4.0  # Optional: For Firebase Auth Provider

# Environment
python-dotenv==1.0.0

# AI Providers
openai==1.3.7
anthropic==0.18.1
tiktoken==0.5.2

# HTTP Client
httpx==0.25.2

# Utilities
python-multipart==0.0.6
email-validator==2.1.0

# Template Engine
Jinja2==3.1.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Code Quality
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1
